+++ tinper-acs/ac-grids/src/RenderColumn.js
@@ -60,16 +60,24 @@ class RenderColumn extends Component {
             textAlign, validate, disabled, 
             options,required,pattern,patternMessage,
             customizeRender,valueField,
-            defaultValue,filedProps,onValidate
+            defaultValue,filedProps,onValidate,record
         } = this.props;
         let placement = 'left';
         if(textAlign)placement=textAlign=='center'?'bottom':textAlign;
         if(customizeRender){
+            let customizeRenderText = customizeRender.type&&customizeRender.type.customizeRenderText;
+            let text = customizeRenderText?customizeRenderText({
+                ...filedProps,
+                value,
+                field: dataIndex,
+                record,
+                index
+            }):value;
             return (<div>
                 {
                     disabled?
-                    <ToolTip overlay={value} inverse placement={placement}>
-                        <span className='ac-grid-cell'>{value}</span>
+                    <ToolTip overlay={text} inverse placement={placement}>
+                        <span className='ac-grid-cell'>{text}</span>
                     </ToolTip>:<RenderCell renderType='refer' text = {value} textAlign={textAlign}>
                     {
                         React.cloneElement(customizeRender,{
@@ -81,6 +89,7 @@ class RenderColumn extends Component {
                             value:value,
                             onValidate:onValidate,
                             index:index,
+                            record:record,
                             onChange:(field, v)=>{this.props.onChange(index,dataIndex,v)},
                             ...filedProps
                         })