+++ tinper-acs/bee-complex-grid/src/RenderColumn.js
@@ -12,6 +12,7 @@ import ToolTip from 'bee-tooltip';
 const propTypes = {
     onChange: PropTypes.func,
     filedProps: PropTypes.object,//filed属性
+    
 }
 
 const defaultProps = {
@@ -48,7 +49,9 @@ class RenderColumn extends Component {
         this.getValue(value);
         this.props.onChange(index, dataIndex, value);
     }
-
+    onRef=(ref)=>{
+        this.customizeRender = ref
+    }
     /**
      * 渲染组件函数
      * @returns JSX
@@ -58,19 +61,56 @@ class RenderColumn extends Component {
             textAlign, validate, disabled,
             required, pattern, patternMessage,
             customizeRender, valueField,
-            filedProps,onValidate,defaultValue
+            filedProps,onValidate,defaultValue,record,forceRenderColumn
         } = this.props;
         let placement = 'left';
         if (textAlign) placement = textAlign == 'center' ? 'bottom' : textAlign;
         if (customizeRender) {
+            let customizeRenderText = this.customizeRender&&this.customizeRender.customizeRenderText;
+            let customText = customizeRenderText&&customizeRenderText({
+                ...filedProps,
+                value,
+                field: dataIndex,
+                record,
+                index
+            });
+            let text = value,overlay = value;
+            if(customText){
+                if(Object.prototype.toString.call(customText)=='[object Object]'){
+                    overlay = customText.overlay;
+                    text = customText.text;
+                }else if(Object.prototype.toString.call(customText)=='[object String]'){
+                    text = customText;
+                    overlay = customText;
+                }
+            }
             return (
                 <div>
                     {
+                        forceRenderColumn?<span style={{'display':'none'}}>
+                        {
+                            React.cloneElement(customizeRender, {
+                                valueField: valueField,
+                                textAlign: textAlign,
+                                field: dataIndex,
+                                validate: validate,
+                                required: required,
+                                value: value,
+                                index:index,
+                                record:record,
+                                onChange: (field, v) => { this.props.onChange(index, dataIndex, v) },
+                                onValidate:onValidate,
+                                onRef:this.onRef,
+                                ...filedProps,
+                            })
+                        }
+                        </span>:''
+                    }
+                    {
                         disabled ?
-                            <ToolTip overlay={value} inverse placement={placement}>
-                                <span className='u-edit-grid-cell'>{value}</span>
-                            </ToolTip>
-                            : <RenderCell type='refer' text={value} textAlign={textAlign}>
+                            <ToolTip overlay={overlay} inverse placement={placement}>
+                                <span className='u-edit-grid-cell'>{text}</span>
+                            </ToolTip>:<RenderCell type='refer' overlay={overlay} text={text} textAlign={textAlign}>
                                 {
                                     React.cloneElement(customizeRender, {
                                         valueField: valueField,
@@ -80,9 +120,11 @@ class RenderColumn extends Component {
                                         required: required,
                                         value: value,
                                         index:index,
+                                        record:record,
                                         onChange: (field, v) => { this.props.onChange(index, dataIndex, v) },
                                         onValidate:onValidate,
-                                        ...filedProps
+                                        onRef:this.onRef,
+                                        ...filedProps,
                                     })
                                 }
                             </RenderCell>